backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "numpy>=2.4.4",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "tensorflow>=2.21.0",
    "uvicorn[standard]>=0.42.0",
]
